refactor: move store config to a separate file
